--- {"sender": "parth", "text": "jweoifjoweijfoiwjef"}
@@ -1,8 +1,0 @@
-HTTP/1.1 200 OK
-Access-Control-Allow-Origin: *
-Content-Type: text/html; charset=utf-8
-ETag: W/"4-hdjXa6Fb3j7xYC9HfzL9ZOMv6lo"
-X-Powered-By: Express
-Date: Tue, 18 Jun 2024 13:32:34 GMT
-Content-Length: 4
-
